+++ web/sites/default/files/civicrm/ext/org.civicoop.civirules/CRM/Civirules/Trigger/Post.php
@@ -18,6 +18,16 @@ class CRM_Civirules_Trigger_Post extends CRM_Civirules_Trigger {
    */
   protected $op;
 
+  public static ?CRM_Civirules_TriggerData_TriggerData $triggerDataCache = NULL;
+
+  public function __construct($trigger = NULL) {
+    if (isset($trigger)) {
+      $this->op = $trigger['op'] ?? NULL;
+      $this->objectName = $trigger['object_name'] ?? NULL;
+    }
+    parent::__construct($trigger);
+  }
+
   /**
    * Getter for object name
    *
@@ -104,12 +114,21 @@ class CRM_Civirules_Trigger_Post extends CRM_Civirules_Trigger {
     if (!in_array($op, $extensionConfig->getValidTriggerOperations())) {
       return;
     }
+    // Delete the cached trigger data in case we modify the same record twice in one process.
+    self::$triggerDataCache = NULL;
 
     // find matching rules for this objectName and op
     $triggers = CRM_Civirules_BAO_CiviRulesRule::findRulesByObjectNameAndOp($objectName, $op);
     foreach($triggers as $trigger) {
       if ($trigger instanceof CRM_Civirules_Trigger_Post) {
+        if (self::$triggerDataCache) {
+          $trigger->setTriggerData(self::$triggerDataCache);
+        }
         $trigger->triggerTrigger($op, $objectName, $objectId, $objectRef, $eventID);
+        // Capture the trigger data for the first trigger so we don't have to query again on future triggers.
+        if ($trigger->hasTriggerData()) {
+          self::$triggerDataCache ??= $trigger->getTriggerData();
+        }
       }
     }
   }
@@ -153,8 +172,10 @@ class CRM_Civirules_Trigger_Post extends CRM_Civirules_Trigger {
       //set also original data with an edit event
       $oldData = CRM_Civirules_Utils_PreData::getPreData($entity, $objectId, $eventID);
       $triggerData = new CRM_Civirules_TriggerData_Edit($entity, $objectId, $data, $oldData);
+      $triggerData->setTrigger($this);
     } else {
       $triggerData = new CRM_Civirules_TriggerData_Post($entity, $objectId, $data);
+      $triggerData->setTrigger($this);
     }
 
     $this->alterTriggerData($triggerData);