+++ web/sites/default/files/civicrm/ext/org.civicoop.civirules/CRM/CivirulesActions/Activity/CreateActivityFromEvent.php
@@ -112,7 +112,7 @@ class CRM_CivirulesActions_Activity_CreateActivityFromEvent extends CRM_Civirule
         'value' => $action_params['status_id'],
         'option_group_id' => 'activity_status',
       ]);
-    } catch (CiviCRM_API3_Exception $e) {
+    } catch (CRM_Core_Exception $e) {
     }
     try {
       $action_params['activity_type_id'] = civicrm_api3('OptionValue', 'getvalue', [
@@ -303,7 +303,7 @@ class CRM_CivirulesActions_Activity_CreateActivityFromEvent extends CRM_Civirule
    * @access public
    */
   public function getExtraDataInputUrl($ruleActionId) {
-    return CRM_Utils_System::url('civicrm/civirule/form/action/activityfromevent', 'rule_action_id='.$ruleActionId);
+    return $this->getFormattedExtraDataInputUrl('civicrm/civirule/form/action/activityfromevent', $ruleActionId);
   }
 
   /**