+++ web/sites/default/files/civicrm/ext/org.civicoop.emailapi/CRM/Emailapi/CivirulesAction/Send.php
@@ -97,7 +97,7 @@ class CRM_Emailapi_CivirulesAction_Send extends CRM_CivirulesActions_Generic_Api
    * $access public
    */
   public function getExtraDataInputUrl($ruleActionId) {
-    return CRM_Utils_System::url('civicrm/civirules/actions/emailapi', 'rule_action_id='.$ruleActionId);
+    return $this->getFormattedExtraDataInputUrl('civicrm/civirules/actions/emailapi', $ruleActionId);
   }
 
   /**
@@ -178,4 +178,55 @@ class CRM_Emailapi_CivirulesAction_Send extends CRM_CivirulesActions_Generic_Api
   public function alterApiParametersForTesting($parameters, CRM_Civirules_TriggerData_TriggerData $triggerData) {
     return $this->alterApiParameters($parameters, $triggerData);
   }
+
+  /**
+   * Get various types of help text for the action:
+   *   - actionDescription: When choosing from a list of actions, explains what the action does.
+   *   - actionDescriptionWithParams: When a action has been configured for a rule provides a
+   *       user friendly description of the action and params (see $this->userFriendlyConditionParams())
+   *   - actionParamsHelp (default): If the action has configurable params, show this help text when configuring
+   * @param string $context
+   *
+   * @return string
+   */
+  public function getHelpText(string $context): string {
+    switch ($context) {
+      case 'actionDescriptionWithParams':
+        return $this->userFriendlyConditionParams();
+
+      case 'actionDescription':
+        return E::ts('Send an email');
+
+      case 'actionParamsHelp':
+        return E::ts('<p>This is the form where you can set what is going to happen with the email.</p>
+    <p>The first few fields are relatively straightforward: the <strong>From Name</strong> is the name the email will be sent from and the <strong>From Email</strong> is the email address the email will be sent from. Leave these blank to use the configured defaults.</p>
+    <p><strong>From Email Addressess</strong> is where you select which From Email Address will be used to send an email. You can create and edit them in <strong>Administer>Communications>FROM Email Addresses</strong></p>
+    <p>The <strong>Message Template</strong> is where you select which CiviCRM message template will be used to compose the mail. You can create and edit them in <strong>Administer>Communications>Message Templates</strong></p>
+    <p>The next section allows you to manipulate where the email will be sent to.<br/>
+    By <strong>default</strong> the email will be sent to the <strong>primary email address of the contact</strong> in question.<br/>
+    You have the option to deviate from that behaviour in 2 ways:
+    <ol>
+      <li>You can select to use an email address of a <strong>specific location type</strong>, for example <em>Work</em>. You do that by selecting the location type. The mail will be sent to the newest email address of that location type (if there are more than one for the contact). If no email address of the specified location type is found, the email will still be sent to the primary email address of the contact</li>
+      <li>If you do not select a location type you can <strong>tick</strong> the <strong>Send to Alternative Email Address</strong> box. You will then get the option to enter an alternative email address. Obviously you have to enter a valid email address here.</li>
+    </ol>
+    </p>
+      <p>Finally you can specify an email address for the <strong>CC to</strong> (a copy of the email will be sent to this email address and the email address will be visible to the recipient of the email too) or the <strong>BCC to</strong> (a copy of the email will be sent to this email address and the email address will NOT be visible to the recipient of the email too).</p>
+      <p>The sending of the email will also lead to an activity (type <em>Email</em>) being recorded for the contact in question, whatever email address will be used.</p>
+    ');
+    }
+
+    return $helpText ?? '';
+  }
+
+  /**
+   * Copy of CiviRules function which did not exist in CiviRules < 3.17
+   * @param string $url
+   * @param int $ruleActionID
+   *
+   * @return string
+   */
+  public function getFormattedExtraDataInputUrl(string $url, int $ruleActionID): string {
+    return CRM_Utils_System::url($url, 'rule_action_id=' . $ruleActionID, FALSE, NULL, FALSE, FALSE, TRUE);
+  }
+
 }