+++ web/sites/default/files/civicrm/ext/org.civicoop.emailapi/CRM/Emailapi/Form/CivirulesAction/SendToCaseContactsOnCase.php
@@ -22,9 +22,7 @@ class CRM_Emailapi_Form_CivirulesAction_SendToCaseContactsOnCase extends CRM_Ema
    */
   public function setDefaultValues() {
     $defaultValues = parent::setDefaultValues();
-    if (!empty($this->ruleAction->action_params)) {
-      $data = unserialize($this->ruleAction->action_params);
-    }
+    $data = $this->unserializeParams();
     if (!empty($data['relationship_type'])) {
       $defaultValues['relationship_type'] = $data['relationship_type'];
     }